Rename from cache to utils
