+++ src/vs/workbench/contrib/positronConnections/browser/components/schemaNavigation.tsx
@@ -8,7 +8,7 @@ import { useStateRef } from 'vs/base/browser/ui/react/useStateRef';
 import { DisposableStore } from 'vs/base/common/lifecycle';
 import { usePositronConnectionsContext } from 'vs/workbench/contrib/positronConnections/browser/positronConnectionsContext';
 import * as DOM from 'vs/base/browser/dom';
-import { IPositronConnectionEntry } from 'vs/workbench/services/positronConnections/browser/positronConnectionsCache';
+import { IPositronConnectionEntry } from 'vs/workbench/services/positronConnections/browser/positronConnectionsUtils';
 import { ActionBar, ACTION_BAR_HEIGHT as kActionBarHeight } from 'vs/workbench/contrib/positronConnections/browser/components/schemaNavigationActionBar';
 import { FixedSizeList as List } from 'react-window';
 import { positronClassNames } from 'vs/base/common/positronUtilities';
